% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindOverlapUsage.R
\name{count_random_5mer_overlaps}
\alias{count_random_5mer_overlaps}
\title{Count exact 5-bp matches of a random motif in a random chromosome}
\usage{
count_random_5mer_overlaps(fasta_path)
}
\arguments{
\item{fasta_path}{Path to a genome FASTA file (DNAStringSet format)}
}
\value{
Integer: number of exact 5-bp matches found
}
\description{
Count exact 5-bp matches of a random motif in a random chromosome
}
\examples{
count_random_5mer_overlaps(
  "C:/Users/fraga/OneDrive/Desktop/Advanced Bioinformatics/Data/
   Homo_sapiens.GRCh38.dna.primary_assembly.fa"
)
}
