% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgRNAExonLocation.R
\name{match_sgrna_to_exons}
\alias{match_sgrna_to_exons}
\title{Match sgRNAs to exons and annotate control values}
\usage{
match_sgrna_to_exons(
  exon_gene_path,
  sgrna_data_path,
  genome_fasta_path,
  output_path
)
}
\arguments{
\item{exon_gene_path}{Path to your exon/gene table (Excel)}

\item{sgrna_data_path}{Path to your sgRNA results (Excel)}

\item{genome_fasta_path}{Path to the genome FASTA file}

\item{output_path}{Path where the annotated sgRNA Excel should be written}
}
\value{
Invisibly returns the final annotated data frame
}
\description{
Match sgRNAs to exons and annotate control values
}
\examples{
\dontrun{
match_sgrna_to_exons(exon_gene_path,sgrna_data_path,genome_fasta_path,output_path)
}
}
