% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_hot_encode.R
\name{one_hot_encode}
\alias{one_hot_encode}
\title{One Hot Encode for DNA}
\usage{
one_hot_encode(sequence)
}
\arguments{
\item{sequence}{}
}
\value{
One Hot Encoded Matrix with position as columns and base as rows
}
\description{
One Hot Encode for DNA
}
\examples{

one_hot_encode("ACGT")

}
